+++ components/migration-client/wso2-api-migration-client/src/main/java/org/wso2/carbon/apimgt/migration/validator/validators/APIValidator.java
@@ -1,3 +1,20 @@
+/*
+ * Copyright (c) 2022, WSO2 LLC. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 package org.wso2.carbon.apimgt.migration.validator.validators;
 
 import org.apache.commons.logging.Log;
@@ -16,6 +33,7 @@ import org.wso2.carbon.apimgt.impl.wsdl.model.WSDLValidationResponse;
 import org.wso2.carbon.apimgt.migration.util.Constants;
 import org.wso2.carbon.apimgt.migration.validator.dao.ApiMgtDAO;
 import org.wso2.carbon.apimgt.migration.validator.utils.Utils;
+import org.wso2.carbon.apimgt.migration.validator.utils.V260Utils;
 import org.wso2.carbon.apimgt.rest.api.publisher.v1.common.mappings.PublisherCommonUtils;
 import org.wso2.carbon.apimgt.rest.api.publisher.v1.dto.GraphQLValidationResponseDTO;
 import org.wso2.carbon.governance.api.exception.GovernanceException;
@@ -23,6 +41,7 @@ import org.wso2.carbon.governance.api.generic.dataobjects.GenericArtifact;
 import org.wso2.carbon.governance.api.util.GovernanceUtils;
 import org.wso2.carbon.registry.core.Resource;
 import org.wso2.carbon.registry.core.exceptions.RegistryException;
+import org.wso2.carbon.registry.core.session.UserRegistry;
 
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
@@ -31,20 +50,66 @@ import java.sql.SQLException;
 import java.util.Set;
 import java.util.regex.Pattern;
 
-public class V410Validator extends Validator {
-    private static final Log log = LogFactory.getLog(V410Validator.class);
+/**
+ * The APIValidator class is used to validate all API related pre-validation steps.
+ * Validations are done on one API and one pre-validation step at a time
+ * using the {@link #validate} method.
+ */
+public class APIValidator {
+    private Utils utils;
+    private UserRegistry registry;
+    private String apiName;
+    private String apiVersion;
+    private String provider;
+    private String apiType;
+    private String apiId;
+    private static final Log log = LogFactory.getLog(APIValidator.class);
     private final String saveSwagger = System.getProperty(Constants.preValidationService.SAVE_INVALID_DEFINITION);
 
-    public V410Validator(Utils utils) {
-        super(utils);
+    public APIValidator(Utils utils) {
+        this.utils = utils;
     }
 
-    @Override
+    /**
+     * @param registry         UserRegistry
+     * @param artifact         Artifact corresponding to the API
+     * @param preMigrationStep Pre-validation step to run
+     * @throws GovernanceException if an error occurs while accessing artifact attributes
+     */
+    public void validate(UserRegistry registry, GenericArtifact artifact, String preMigrationStep)
+            throws GovernanceException {
+        this.registry = registry;
+        this.apiName = artifact.getAttribute(APIConstants.API_OVERVIEW_NAME);
+        this.apiVersion = artifact.getAttribute(APIConstants.API_OVERVIEW_VERSION);
+        this.provider = artifact.getAttribute(APIConstants.API_OVERVIEW_PROVIDER);
+
+        // At this point of  pre-validation step, SOAP and SOAPTOREST APIs from 2.6.0 will have their overview_type
+        // set as HTTP, hence we are employing a Util method to fetch correct API Type based on other resources and
+        // artifact fields.
+        if (Constants.VERSION_2_6_0.equals(utils.getMigrateFromVersion())) {
+            this.apiType = V260Utils.getAPIType(artifact);
+        } else {
+            this.apiType = artifact.getAttribute(APIConstants.API_OVERVIEW_TYPE);
+        }
+
+        this.apiId = artifact.getId();
+        if (Constants.preValidationService.API_ENDPOINT_VALIDATION.equals(preMigrationStep)) {
+            validateEndpoints();
+        } else if (Constants.preValidationService.API_DEFINITION_VALIDATION.equals(preMigrationStep)) {
+            validateAPIDefinition();
+        } else if (Constants.preValidationService.API_AVAILABILITY_VALIDATION.equals(preMigrationStep)) {
+            validateApiAvailability();
+        } else if (Constants.preValidationService.API_RESOURCE_LEVEL_AUTH_SCHEME_VALIDATION.equals(preMigrationStep)) {
+            validateApiResourceLevelAuthScheme();
+        } else if (Constants.preValidationService.API_DEPLOYED_GATEWAY_TYPE_VALIDATION.equals(preMigrationStep)) {
+            validateApiDeployedGatewayType(artifact);
+        }
+    }
+
     public void validateEndpoints() {
 
     }
 
-    @Override
     public void validateAPIDefinition() {
         if (!utils.isStreamingAPI(apiType)) {
             validateOpenAPIDefinition();
@@ -59,7 +124,6 @@ public class V410Validator extends Validator {
         }
     }
 
-    @Override
     public void validateApiAvailability() {
         try {
             log.info("Validating API availability in db for API {name: " + apiName + ", version: " +
@@ -216,7 +280,6 @@ public class V410Validator extends Validator {
         }
     }
 
-    @Override
     public void validateApiResourceLevelAuthScheme() {
 
         Pattern pattern = Pattern.compile("2\\.\\d\\.\\d");
@@ -246,7 +309,6 @@ public class V410Validator extends Validator {
 
     }
 
-    @Override
     public void validateApiDeployedGatewayType(GenericArtifact apiArtifact) {
         log.info("Validating deployed gateway type for API {name: " + apiName + ", version: " + apiVersion
                 + ", provider: " + provider + "}");
